Reorganised folders (#46)
* moved toggle

* removed redundant

* moved modals

* moved panel

* moved button

* moved other

* renamed foldes

* reverted defaults

* fixed tests
