+++ src/components/Main/Greetings.test.jsx
@@ -1,5 +1,5 @@
 import Greetings from './Greetings'
-import { render, screen } from '../utils/test-utils'
+import { render, screen } from '../../utils/test-utils'
 
 describe('greeting in the document', () => {
   it('text is visible', () => {
